Refactor: 아티클 서비스 마이그레이션
